# Conflicts:
#	Programs/TestBed/Sources/Tests/NotificationTest.cpp


Former-commit-id: 0cd424f280be6c3916603ba89e28faec6aea860f

+++ Sources/Internal/Engine/Private/Qt/RenderWidgetOGL.cpp
@@ -224,7 +224,7 @@ void RenderWidgetOGL::InitCustomRenderParams(rhi::InitParam& params)
     params.threadedRenderFrameCount = 1;
     params.acquireContextFunc = &AcquireContextImpl;
     params.releaseContextFunc = &ReleaseContextImpl;
-    params.defaultFrameBuffer = reinterpret_cast<void*>(quickWindow()->renderTarget()->handle());
+    params.defaultFrameBuffer = reinterpret_cast<void*>(static_cast<intptr_t>(quickWindow()->renderTarget()->handle()));
 }
 
 void RenderWidgetOGL::AcquireContext()
